requirements.txt
asgiref==3.2.10
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
django-baton==1.9.3
django-crispy-forms==1.9.2
django-jalali-date==0.3.1
django-summernote==0.8.11.6
django-widget-tweaks==1.4.8
google-api-core==1.21.0
google-api-python-client==1.10.0
google-auth==1.19.2
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
jdatetime==3.6.2
oauth2client==1.5.2
Pillow==7.2.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
